Fixed deadlock if the SharedScan consumer is on slice 0.
This happend because in the case of slice 0, sending a notification about the
completion of reading will occur only after receiving the result of query, while
the writing part will wait for the notification and will not shutdown. Which
leads to that slice 0 will never get a result.

This patch fixes this by compiling a list of consumers for the current slide and
sending notifications before recive result.