Filter excessive attrs from ORCA's DML delete node
Currently, ORCA passes all attributes to DML node regardless of the type of
operation. Passing all user-defined attrs to delete operation causing
unnecessary scans of this attrs. This may lead to performance decreasing,
especially for column-oriented tables. In the most of cases such attrs may be
omitted. The only exception is the "after" trigger, which needs all attrs to
work.
The goal of this patch is to change the behavior of delete operation and use
empty source array for DML node if possible. To use it, execution part of DML
node was also changed. From now, it can work with subset of attrs in target
list, which is close to how postgres works.