Fix warnings in Jumbling logic
Problem description:
Following warning messages appear
    "WARNING: unrecognized node type: 925"
    "WARNING: unrecognized node type: 924"
if 'pg_stat_statements' extension is enabled and query,
containing GROUP BY clause or GROUPING function, is executed.

Cause:
In JumbleExpr, called from JumbleQuery, there is no handling
for node tags T_GroupingClause and T_GroupingFunc.

Fix:
Added handling logic for missed node tags.
    For T_GroupingFunc we append to the jumble the tag itself
and list of arguments. Field 'ngrpcols' is not appended to the
jumble, because it can be deduced from the list of arguments
in grouping clause. Plus added handling for T_Integer tag,
because it is required to parse list of arguments.
    For T_GroupingClause we append to the jumble the tag itself,
grouping type and list of groupsets. Field 'location' is not
appended to the jumble, because it is the textual location from
parser and is not semantically significant.