Improve test for pg_stat_statements
Modified the pg_stat_statements test to enable the pg_stat_statements
extension automatically at the beginning of the test, and disable the
extension once the test is complete.
Also added minor refactoring of recently added code in JumbleExpr - moved
all new code to the end of the switch case in order to simplify the patch.